requirements.txt
litellm>=1.40.0
pyyaml>=6.0
typer>=0.9.0
rich>=13.0
PyGithub>=2.0
GitPython>=3.1
langgraph>=0.2.0
python-dotenv>=1.0.0
